Improve expgfx_renderSourcePools 95.5->100% (u8 cast on frustum return, #11)

+++ src/main/dll/dll_000A_expgfx.c
@@ -2009,7 +2009,7 @@ void expgfx_renderSourcePools(int sourceId, int sourceMode)
             (*poolSourceModes == sourceMode + EXPGFX_POOL_SOURCE_MODE_SOURCE_OFFSET))
         {
             boundsTemplate = Expgfx_GetBoundsTemplate(*poolBoundsTemplateIds);
-            if (frustumTestAabbWithPlaneOffsets(poolBounds->minX - playerMapOffsetX,
+            if ((u8)frustumTestAabbWithPlaneOffsets(poolBounds->minX - playerMapOffsetX,
                                                 poolBounds->maxX - playerMapOffsetX,
                                                 poolBounds->minY, poolBounds->maxY,
                                                 poolBounds->minZ - playerMapOffsetZ,
